lib: move ExpiringModel to lib from common
