+++ authentik/stages/invitation/models.py
@@ -8,9 +8,9 @@ from django.utils.translation import gettext_lazy as _
 from django.views import View
 from rest_framework.serializers import BaseSerializer, Serializer
 
-from authentik.core.models import ExpiringModel, User
+from authentik.core.models import User
 from authentik.flows.models import Stage
-from authentik.lib.models import SerializerModel
+from authentik.lib.models import ExpiringModel, SerializerModel
 
 
 class InvitationStage(Stage):
